--- libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/navigation/NavigationRouteEventListener.kt
@@ -1,22 +1,0 @@
-package com.mapbox.services.android.navigation.v5.navigation
-
-import com.mapbox.navigation.utils.time.ElapsedTime
-import okhttp3.Call
-import okhttp3.EventListener
-
-class NavigationRouteEventListener
-@JvmOverloads
-constructor(
-    val time: ElapsedTime = ElapsedTime()
-) : EventListener() {
-
-    override fun callStart(call: Call) {
-        super.callStart(call)
-        time.start()
-    }
-
-    override fun callEnd(call: Call) {
-        super.callEnd(call)
-        time.end()
-    }
-}
