+++ libdirections-offboard/build.gradle
@@ -22,6 +22,8 @@ dependencies {
     implementation project(':libnavigation-util')
     api(project(':libnavigation-base'))
 
+    api(project(':libnavigation-base'))
+
     // Support
     implementation dependenciesList.supportAnnotation
