+++ libdirections-offboard/src/main/java/com/mapbox/navigation/route/offboard/MapboxOffboardRouter.kt
@@ -49,7 +49,6 @@ class MapboxOffboardRouter(
     }
 
     override fun getRoute(origin: Point, waypoints: List<Point>, listener: Router.RouteListener) {
-
     }
 
     override fun cancel() {
